+++ src/main/java/btw/lowercase/skyboxify/SkyboxifyInfo.java
@@ -1,0 +1,7 @@
+package btw.lowercase.skyboxify;
+
+public final class SkyboxifyInfo {
+	public static final String MOD_ID = "@MODID@";
+	public static final String VERSION = "@VERSION@";
+	public static final String COMMIT = "@COMMIT@";
+}